Proxy is now more generic
